Cleaned up file output strucutre in gep_analysis.py and finally got gep_wcd_launch.py to call grd_evaluate

+++ src/gep_wcd_launch.py
@@ -6,59 +6,80 @@ Launch WCD for a all problems of a domain and save output in GEP locations
 @author: adamab
 '''
 
-import subprocess
 import os
 import time
 import shutil
 import sys
-#import grd_evaluator
+import grd_evaluator
+import grd_options
 
 def gep_wcd_launch(grd_dir, domain_name):
 
+    #setup filenames and directories for output and copying
     gep_data_path = os.path.join(*[grd_dir, "gep", "data", domain_name])
-    #domain_path = os.path.join(*[grd_dir,"benchmarks", domain_name])
     gen_files_path = os.path.join(grd_dir, "gen_files")
     log_files_path = os.path.join(grd_dir, "log_files")
-
-    problem_dirs = [x[1] for x in os.walk(gep_data_path)][0]
-
-    #problem_dirs = ['p01', 'p02', 'p03', 'p04', 'p05']
     wcd_results_filename = "grd_log_reduction.txt"
     wcd_results_filepath = os.path.join(log_files_path, wcd_results_filename)
 
+    #exit if we can't find the GEP destination for all WCD output
+    if (not(os.path.exists(gep_data_path))):
+        print("Exiting, can't find: " + gep_data_path)
+        sys.exit(0)
+
+    #get all the problems in the domain's directory
+    problem_dirs = [x[1] for x in os.walk(gep_data_path)][0]
+    print("Running grd_evaluator_reduce for domain: " + domain_name + " and problems: " + ", ".join(problem_dirs))
+
+    #loop through each problem and call run the GRD process for WCD
     for problem_dir in problem_dirs:
-	#print(problem_dirs)
         problem_name = os.path.basename(problem_dir).split(".")[0]
         problem_path = os.path.join(gep_data_path, problem_name)
-
-        #TODO this block can probably be removed
-	#execute string
-        exec_string = "./scripts/grd_evaluator_reduce " + os.path.join(gep_data_path, "domain.pddl") +" "+ os.path.join(problem_path, "template.pddl") + " " + os.path.join(problem_path, "hyps.dat") + " LatestSplit 1 NA NA True max NA"
-	#grd_evaluator.main(["grd_evaluator_reduce", os.path.join(gep_data_path, "domain.pddl"), os.path.join(problem_path, "template.pddl"), os.path.join(problem_path, "hyps.dat"), "LatestSplit", "1", "NA", "NA", "True", "max", "NA"])
-        #exec_string = ["grd_evaluator_reduce", os.path.join(gep_data_path, "domain.pddl"), os.path.join(problem_path, "template.pddl"), os.path.join(problem_path, "hyps.dat"), "LatestSplit", "1", "NA", "NA", "True", "max", "NA"]
-        print(exec_string)
-        #subprocess.call(exec_string,cwd=grd_dir+"/scripts")
         
-	#TODO below doesn't work and issue #1 on this repo refers to the error from .main() below
-	#TODO also check the harded coded params below to make sure they are correct, we may want "all_pairs" instead of "max" for -v
-        #python3 ./src/grd_evaluator.py -o $1 -p $2 -y $3 -c $4 -g $5 -b $6 -u $7 -f $8 -v $9 -a ${10}
-        exec_array = ["WcdReduce", "-o", os.path.join(gep_data_path, "domain.pddl"), "-p", os.path.join(problem_path, "template.pddl"), "-y", os.path.join(problem_path, "hyps.dat"), "-c", "LatestSplit", "-g", "1", "-b", "NA", "-u", "NA", "-f", "True", "-v", "max", "-a", "NA"]
-        grd_evaluator.main(exec_array)
+	#TODO check the harded coded params below to make sure they are correct, we may want "all_pairs" instead of "max" for -v
+        exec_array = [ "-o", os.path.join(gep_data_path, "domain.pddl"), "-p", os.path.join(problem_path, "template.pddl"), "-y", os.path.join(problem_path, "hyps.dat"), "-c", "WcdReduce-LatestSplit", "-g", "1", "-b", "NA", "-u", "NA", "-f", "True", "-v", "all_pairs", "-a", "NA"]
+
+	#For easy reference above
+	#print ( "-e  --experiment <file or folder>  Plan Recognition experiment files (tar'ed)",file =  sys.stderr)
+	#print ( "-d  --destination folder (the subfolder of pgrd_defs.gen_files where the files are generated)",file = sys.stderr )
+	#print ( "-c  --calculation method    the method to be applied",file =  sys.stderr)
+	#print ( "-v  --combinations examined - either max over all pairs or the wcd of each pair",file =  sys.stderr)
+	#print ( "-h  --help                       Get Help",file =  sys.stderr)
+	#print ( "-t  --max-time <time>            Maximum allowed execution time (defaults to 1800 secs)",file =  sys.stderr)
+	#print ( "-m  --max-memory <time>          Maximum allowed memory consumption (defaults to 1Gb)",file =  sys.stderr)
+	#print ( "-x  --done-file-name          task names that need not be calculated",file =  sys.stderr)
+	#print ( "-o  --domain-file-name         grd domain file name(when specifing files instead of folder)",file =  sys.stderr)
+	#print ( "-p  --temPlate-file-name        grd template file name(when specifing files instead of folder)",file =  sys.stderr)
+	#print ( "-y  --hyps-file-name        grd hyps file name(when specifing files instead of folder)",file =  sys.stderr)
+	#print ( "-b  --obs-file-name        grd non-observable actions file name(when specifing files instead of folder)",file =  sys.stderr)
+	#print ( "-a  --action-tokens-file-name      grd action token file name(when specifing files instead of folder)",file =  sys.stderr)
+	#print ( "-u  --sub-optimal-bound-array",file =  sys.stderr)
+	#print ( "-f  --delete-log-folders  should the log folders be deleted",file =  sys.stderr)
+	#print ( "-r  --rec-obs-file-name     reciproce-observability actions file name(when specifing files instead of folder)",file =  sys.stderr)
+	#print ( "-g  --design-budget-array",file =  sys.stderr)
+	#print ( "-k  --token-file-string",file =  sys.stderr)
+	#print ( "-i  --actions-to-remove",file =  sys.stderr)
+	#print ( "-z  --actions-constraints",file =  sys.stderr)
+	#print ( "-n  --domain",file = sys.stderr)
+        
+        #Error when calling main in grd_evaluator, so just running the contents
+        #grd_evaluator.main(exec_array)
+        options = grd_options.Program_Options( exec_array )
+        grd_evaluator.evaluate(options)
 
         #move results and rename with timestamp
         timestr = time.strftime("%Y%m%d_%H%M%S")
-	#dest_filepath = os.path.join(*[gep_data_path, domain_name, problem_name])
         shutil.copyfile(wcd_results_filepath, os.path.join(problem_path, timestr + "_" + wcd_results_filename) )
         shutil.make_archive(os.path.join(problem_path, timestr + "_gen"), 'zip', gen_files_path)
         shutil.make_archive(os.path.join(problem_path, timestr + "_log"), 'zip', log_files_path)
 
-if __name__=="__main__": 
+if __name__=="__main__":
+
     if len(sys.argv) < 3:
         print("Usage: gep_wcd_launch grd_path domain_name")
         sys.exit()
     else:
         grd_path = sys.argv[1]
         domain_name = sys.argv[2]
-#         grd_path = "\home\aamosbinks\git\goal-recognition-design"
-#         domain_name = "easy-grid"
+
         gep_wcd_launch(grd_path, domain_name)
